# parallelize `getTableStats` when `show accounts`