## What type of PR is this?

- [ ] API-change
- [ ] BUG
- [x] Improvement
- [ ] Documentation
- [ ] Feature
- [ ] Test and CI
- [ ] Code Refactoring

## Which issue(s) this PR fixes:

issue #https://github.com/matrixorigin/MO-Cloud/issues/2564

## What this PR does / why we need it:
one `getTableStats` execution may take 10ms to 100ms, and `show accounts` will spend 1s to10s with 100 accounts. this is even worse when MO has been high pressure.

this PR parallelized `getTableStats` making it 3x~5x faster than before.

however, `show accounts` still could exceed 30 seconds if there are more accounts and a high load.

cherry-pick to 1.1: https://github.com/matrixorigin/matrixone/pull/14718